Fix bug causing type error on game reset
While Game.startGame() worked when FIRST called by a click on the "Enter" button, it threw a type error when a pre-existing game had already been started and the user clicked "Enter" after quitting or "New Game" after a defeat.

The error message read: "Uncaught TypeError: this.gameMap.rooms[this.player.currentLocation].neighbors.forEach is not a function". Here, "this" refers to the current instance of the Game object, and gameMap.rooms is a property containing an array of "room" objects. Each "room" object contains a "neighbors" property, each supposedly containing an array representing the ID's of neighboring rooms. The "forEach" function gets called in another method nested inside the startGame() method, "game.gameMap.populateGameMap()". Its purpose is to initialize a new Room object for each array item contained in gameMap.rooms, mutating the array items as it does so (not replacing them).

The problem is that the Room constructor function creates a "neighbors" property by referencing the original contents of the gameMap.rooms array before it gets mutated. It does this because when the js script is first loaded, each gameMap.rooms array item contains not an object but rather a smaller array representing the ID's of neighboring rooms. In other words, every new Room object, when initialized, "absorbs" or integrates the information contained in the array slot it will eventually occupy.

So: When the user clicks "Enter" when FIRST starting a game, each new Room object integrates an array into its "neighbors" property. When the user clicks "Enter" after quitting or "New Game" after a defeat, however, gameMap.rooms already contains an array of Room objects. Hence, ".forEach" is running on an object rather than an array, and hence the type error.

I fixed the error by moving the initializer for a new GameMap object from the Game class constructor function into the Game.startGame() method, before the populateGameMap() method runs. This serves to reset the array items in gameMap.rooms to their original state as arrays, not objects.